+++ util-core/src/main/scala/com/twitter/conversions/U64Ops.scala
@@ -8,14 +8,14 @@ object U64Ops {
    *
    * @see [[java.lang.Long.parseUnsignedLong()]]
    */
-  implicit class StringOps(val self: String) extends AnyVal {
+  implicit class RichU64String(val self: String) extends AnyVal {
     def toU64Long: Long = java.lang.Long.parseUnsignedLong(self, 16)
   }
 
   /**
    * Converts this unsigned 64-bit long value into a 16-character HEX string.
    */
-  implicit class LongOps(val self: Long) extends AnyVal {
+  implicit class RichU64Long(val self: Long) extends AnyVal {
     def toU64HexString: String = "%016x".format(self)
   }